+++ src/pages/MyPage.jsx
@@ -4,7 +4,8 @@ import icon_pencil from '@/assets/icons/icon_pencil.png'
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { v4 as uuidv4 } from 'uuid'
-import { getPreSignedUrl } from '../api/albums/presignedUrl'
+
+import { getPreSignedUrl } from '../api/album'
 import { updateUserInfo } from '../api/user/userInfoUpdate'
 import ConfirmModal from '../components/common/ConfirmModal'
 import Header from '../components/common/Header'
